requirements.txt
face_recognition
ttkbootstrap
Pillow
tqdm
````